Исправляет добавление карты
